Gjort om sendBrevSchema.js til TS
